backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# Core research dependencies
langgraph>=0.5.4
langchain-community>=0.3.9
langchain-openai>=0.3.28
langchain-anthropic>=0.3.15
langchain-mcp-adapters>=0.1.6
langchain-tavily
langchain-google-genai>=2.1.5
tavily-python>=0.5.0
python-dotenv>=1.0.1
httpx>=0.24.0
markdownify>=0.11.6
beautifulsoup4==4.13.3
requests>=2.32.3
rich>=13.0.0
mcp>=1.9.4
langsmith>=0.3.37
